Updated: Change package to khttp
The most breaking of breaking changes. Brevity is an important thing!
