--- src/main/kotlin/me/kyleclemens/khttp/structures/authorization/Authorization.kt
@@ -1,12 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package me.kyleclemens.khttp.structures.authorization
-
-interface Authorization {
-
-    val header: Pair<String, String>
-
-}
