+++ src/test/kotlin/khttp/helpers/StringIterable.kt
@@ -1,0 +1,10 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package khttp.helpers
+
+class StringIterable(val string: String) : Iterable<Char> {
+    override fun iterator(): Iterator<Char> = this.string.toCharArray().iterator()
+}
