--- src/test/kotlin/me/kyleclemens/khttp/KHttpPatchSpec.kt
@@ -1,23 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package me.kyleclemens.khttp
-
-import kotlin.test.assertEquals
-
-class KHttpPatchSpec : MavenSpek() {
-    override fun test() {
-        given("a patch request") {
-            val url = "https://httpbin.org/patch"
-            val request = patch(url)
-            on("accessing the json") {
-                val json = request.jsonObject
-                it("should have the same url") {
-                    assertEquals(url, json.getString("url"))
-                }
-            }
-        }
-    }
-}
