--- src/test/kotlin/me/kyleclemens/khttp/helpers/StringIterable.kt
@@ -1,10 +1,0 @@
-/*
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-package me.kyleclemens.khttp.helpers
-
-class StringIterable(val string: String) : Iterable<Char> {
-    override fun iterator(): Iterator<Char> = this.string.toCharArray().iterator()
-}
